+++ covmatic_lws/check_update.py
@@ -4,6 +4,9 @@ from typing import Tuple, Iterable, Callable
 from itertools import zip_longest
 from functools import partialmethod
 import operator
+import logging
+
+logger = logging.getLogger(__name__)
 
 
 class Version(str):
@@ -47,16 +50,23 @@ def latest_version(pkg_name: str = "covmatic-localwebserver", index=None, python
             return s.split("(")[1].split(")")[0]
     return __version__
 
+# TODO: re-enable the version check service.
+#       now it takes some second to say that the service has been disabled due to unmanagabe load;
+#       so we disable it because it is useless.
+# def up_to_date(current_version: str = __version__, pkg_name: str = "covmatic-localwebserver", index=None, python_lookup: str = os.sys.executable, python_install: str = None) -> Tuple[bool, str, str, str]:
+#     """Returns:
+#         up_to_date: whether the package is up to date or not
+#         current_version: the current version
+#         latest_version: the latest version on PyPI
+#         update_command: the command to update the package
+#     """
+#     if python_install is None:
+#         python_install = python_lookup
+#     lv = latest_version(pkg_name=pkg_name, index=index and os.path.dirname(index), python=python_lookup)
+#     cmd = "{} -m pip install{} --upgrade {}".format(python_install, " -i {}".format(index) if index else "", pkg_name)
+#     return Version(lv) <= current_version, current_version, lv, cmd
 
 def up_to_date(current_version: str = __version__, pkg_name: str = "covmatic-localwebserver", index=None, python_lookup: str = os.sys.executable, python_install: str = None) -> Tuple[bool, str, str, str]:
-    """Returns:
-        up_to_date: whether the package is up to date or not
-        current_version: the current version
-        latest_version: the latest version on PyPI
-        update_command: the command to update the package
-    """
-    if python_install is None:
-        python_install = python_lookup
-    lv = latest_version(pkg_name=pkg_name, index=index and os.path.dirname(index), python=python_lookup)
-    cmd = "{} -m pip install{} --upgrade {}".format(python_install, " -i {}".format(index) if index else "", pkg_name)
-    return Version(lv) <= current_version, current_version, lv, cmd
+    logger.info("Update check disabled; please check if XMLRPC PyPI sevice will be re-enabled. "
+                "Go to https://status.python.org/ for more information.")
+    return True, "", "", ""